chore: open drawer for change history
