+++ frontend/src/components/ChangeHistory/index.ts
@@ -1,5 +1,11 @@
 import AffectedTableSelect from "./AffectedTableSelect.vue";
+import ChangeHistoryDetail from "./ChangeHistoryDetail.vue";
 import ChangeHistoryStatusIcon from "./ChangeHistoryStatusIcon.vue";
 import ChangeHistoryTable from "./ChangeHistoryTable.vue";
 
-export { ChangeHistoryTable, ChangeHistoryStatusIcon, AffectedTableSelect };
+export {
+  ChangeHistoryTable,
+  ChangeHistoryDetail,
+  ChangeHistoryStatusIcon,
+  AffectedTableSelect,
+};
